build: :wrench: build package using rollup and export types
