+++ packages/create-navigation-span/package.json
@@ -4,9 +4,16 @@
   "private": true,
   "version": "2.10.0",
   "license": "MIT",
-  "main": "lib/index.ts",
-  "files": [],
-  "dependencies": {
+  "type": "module",
+  "main": "dist/index.js",
+  "files": ["dist"],
+  "types": "dist/types/index.d.ts",
+  "scripts": {
+    "build": "rollup --config",
+    "build:cdn": "npm run build",
+    "clean": "rm -rf dist/*" 
+  },
+  "devDependencies": {
     "@bugsnag/core-performance": "*"
   }
 }
