+++ packages/create-navigation-span/rollup.config.mjs
@@ -1,0 +1,3 @@
+import createRollupConfig from '../../.rollup/index.mjs'
+
+export default createRollupConfig()
